requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
idna==2.8
multi-key-dict==2.0.3
pbr==5.1.1
PyMySQL==0.9.3
python-jenkins==1.4.0
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
